Fixed destination folder in deploy.sh. Fixed #20.

+++ check-build.sh
@@ -59,7 +59,9 @@ prepend-path LD_LIBRARY_PATH    $::env(CLASSPATH)
 MODULE_FILE
 ) > modules/${VERSION}
 mkdir -p ${LIBRARIES_MODULES}/${NAME}
-cp modules/${VERSION} ${LIBRARIES_MODULES}/${NAME}
+if [ ! -f ${LIBRARIES_MODULES}/${NAME} ]; then
+  cp modules/${VERSION} ${LIBRARIES_MODULES}/${NAME}
+fi
 
 echo "[check-build.sh] - Checking REPAST module"
 module add $NAME/$VERSION
